architectural and design decisions, new card-link component
