+++ src/components/main-container/main-container.styles.js
@@ -9,25 +9,6 @@ export const CatsContainer = styled.div`
   height: 100%;
 `;
 
-export const ShuffleButton = styled.button`
-  font-size: 2em;
-  border-radius: 30px;
-  background-color: orange;
-  font-family: 'Chango', cursive;
-  margin: 1em;
-  cursor: pointer;
-  min-width: 200px;
-`;
-
-export const CatsContentWrapper = styled.div`
-  width: 60%;
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin: 0 auto;
-  height: 90%;
-`;
-
 export const CatsTitle = styled.div`
   font-size: 3em;
   color: #0ea5c8;
